LA - all layers second edition (folders reorganization)
