+++ controller/build.gradle
@@ -6,5 +6,4 @@ plugins {
 dependencies{
 
     implementation project(':service')
-    implementation project(':service-impl')
 }
