+++ controller/src/main/java/com/epam/esm/controller/Controller.java
@@ -1,6 +1,6 @@
 package com.epam.esm.controller;
 
-import com.epam.esm.GCService;
+import com.epam.esm.service.GCService;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RestController;
