--- dao-impl/build.gradle
@@ -1,8 +1,0 @@
-plugins {
-    id 'java'
-}
-
-
-dependencies {
-    implementation project(':dao')
-}
