+++ dao/src/main/java/com/epam/esm/dao/DAO.java
@@ -1,4 +1,4 @@
-package com.epam.esm;
+package com.epam.esm.dao;
 
 public interface DAO {
