+++ dao/src/main/java/com/epam/esm/dao/DAOImpl.java
@@ -1,5 +1,6 @@
-package com.epam.esm;
+package com.epam.esm.dao;
 
+import com.epam.esm.dao.DAO;
 import org.springframework.stereotype.Repository;
 
 @Repository
