--- service-impl/build.gradle
@@ -1,9 +1,0 @@
-plugins {
-    id 'java'
-}
-
-dependencies{
-    implementation project(':service')
-    implementation project(':dao')
-    implementation project(':dao-impl')
-}
