+++ service/build.gradle
@@ -1,19 +1,4 @@
-plugins {
-    id 'java'
-}
+dependencies{
 
-group 'by.epam.esm'
-version '1.0-SNAPSHOT'
-
-repositories {
-    mavenCentral()
-}
-
-dependencies {
-    testImplementation 'org.junit.jupiter:junit-jupiter-api:5.7.0'
-    testRuntimeOnly 'org.junit.jupiter:junit-jupiter-engine:5.7.0'
-}
-
-test {
-    useJUnitPlatform()
+    implementation project(':dao')
 }
